Temporary - Disable S3 MD5 Validation
when uploading the index.m3u8 key, MD5 generation is not working
properly.

I’ve tried manually setting the MD5 via PutObjectRequest#setMetaData,
but somehow the metadata is null when received by AmazonS3Client.

It appears that on subsequent uploads of the index.m3u8 file, the MD5
of the original file is used. On the Android filesystem, each m3u8 is
at a unique location, though the AWS S3 key of ‘…/index.m3u8’ is shared…