2.0 foundation
Overhaul of the API client founded on Retrofit and OkHttp with a set of
reactive method alternatives to the callback based originals.

Goal is to significantly remove opportunity for receiving an API client
in an odd state. The old api would return the API client instance
before it was fully configured, forcing the user to observe (or ignore
:) ) callbacks indicating it’s status. Now, the api client is not
returned to the client developer until it has completed the OAuth flow
and is initialized with a Kickflip user. This way you have a guarantee
that when the library provides you with the client you’re good to go.

Also address multiple bugs of the old API client including improperly
storing users in certain cases, and improperly refreshing OAuth tokens.